add homeWork_1 (задания HUman, Hamburger)
