+++ LECTURES/2/index.html
@@ -54,6 +54,7 @@
 
         let oleg = new Male('Oleg', 28);
         let olga = new Female('Olga', 25);
+        
         //1) сделать так, чтобы метод goWork приносил человеку доход, определяемый его "особым" ежечасным доходом * на раб время за 1 рабочий "день"
         //2) дать возможность персонажам "связываться узами"
         //3) запрещать делать покупки при отр. балансе
