+++ Students/Alexey Turdakov/project/0/public/index.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../src/layout/css/style.css">
+    <link rel="stylesheet" href="../src/layout/css/normalize.css">
+    <title>Document</title>
+</head>
+<body>
+    <header>
+        <div class="logo">GeekShop</div>
+        <div class="basket">
+            <form action="#" class="search-form">
+                <input type="text" class="search-field">
+                <button class="btn-search">
+                    <i class="fas fa-search"></i>
+                </button>
+            </form>
+            <button class="btn-basket">Корзина</button>
+            <div class="basket-block invisible">
+                <div class="basket-items">
+                    <!--Basket elems-->
+                    
+                </div>
+            </div>
+        </div>
+    </header>
+    <main>
+        <div class="catalog-items">
+            <!--Catalog elems-->
+            
+        </div>
+    </main>
+    
+    <script defer src="https://use.fontawesome.com/releases/v5.7.2/js/all.js" integrity="sha384-0pzryjIRos8mFBWMzSSZApWtPl/5++eIfzYmTgBBmXYdhvxPc+XcFEk+zJwDgWbP" crossorigin="anonymous"></script>
+    <script src="../src/components/main.js"></script>
+</body>
+</html>
