+++ Students/Alexey Turdakov/project/1/src/index.js
@@ -1,0 +1,7 @@
+import './layout/css/style.css'
+import './layout/css/normalize.css'
+
+import app from './components/main.js'
+
+app();
+
