+++ Students/Alexey Turdakov/project/1/webpack.config.js
@@ -1,0 +1,24 @@
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+    
+    plugins: [new MiniCssExtractPlugin({
+            filename: 'style/[name].css',
+            chunkFilename: '[id].css',
+        }),
+        new HtmlWebpackPlugin({
+            template: './public/index.html'
+        })
+    ],
+    module: {
+        rules: [{
+            test: /\.css$/i,
+            use: [MiniCssExtractPlugin.loader, 'css-loader'],
+        }, ],
+    },
+    devServer: {
+        port: 3000,
+        open: false,
+    }
+};
